requirements.txt
beautifulsoup4==4.7.1
blis==0.2.4
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
cymem==2.0.2
Django==2.1.7
ffmpeg-python==0.1.17
future==0.17.1
google-api-core==1.8.2
google-auth==1.6.3
google-cloud-vision==0.36.0
googleapis-common-protos==1.5.9
grpcio==1.19.0
idna==2.8
jsonschema==2.6.0
murmurhash==1.0.2
numpy==1.16.2
plac==0.9.6
preshed==2.0.1
protobuf==3.7.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pytube==9.4.0
pytz==2018.9
requests==2.21.0
rsa==4.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
sklearn==0.0
soupsieve==1.8
spacy==2.1.2
srsly==0.0.5
thinc==7.0.4
tqdm==4.31.1
urllib3==1.24.1
wasabi==0.2.0
wikipedia==1.4.0
youtube-transcript-api==0.1.3
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.1.0/en_core_web_sm-2.1.0.tar.gz
